Split contactql into a light AST/evaluation package and a parse package
The contactql package keeps the AST types, Resolver, attribute definitions,
query validation, evaluation, inspection and errors — and no longer depends
on the ANTLR runtime. Parsing moves to a new contactql/parse package
(parse.Query) which holds the ANTLR-generated parser glue and visitor.

New light-package constructors let parsing live outside: NewContactQuery
(validates and simplifies a root node) and NewImplicitCondition (interprets
a bare literal as a URN, phone number or name condition). QueryError.WithExtra
and the Attributes map are now exported for the parse package.
